Email/requirements.txt
aiosmtplib==3.0.2
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
billiard==4.2.1
blinker==1.9.0
celery==5.5.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.12
fastapi-mail==1.5.0
h11==0.16.0
idna==3.10
Jinja2==3.1.6
kombu==5.5.3
MarkupSafe==3.0.2
prompt_toolkit==3.0.51
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
redis==6.1.0
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
uvicorn==0.34.2
vine==5.1.0
wcwidth==0.2.13
